# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SINAN AY - YUKSEK LISANS TEST OTOMASYON API",
    "description": "Bu API, Sinan Ay Yüksek Lisans Bitirme Tezi projesi olan test otomasyonu için geliştirilmiştir.",
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "localhost:4002"
    }
  ],
  "tags": [
    {
      "name": "Account",
      "description": "Kullanıcı yetkilendirme işlemleri"
    },
    {
      "name": "Customers",
      "description": "Müşteri İşlemleri"
    }
  ],
  "paths": {
    "/account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Giriş",
        "description": "Giriş",
        "operationId": "loginRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Kayıt Ol",
        "description": "Kayıt Ol",
        "operationId": "registerRequest",
        "requestBody": {
          "description": "Update an existent pet in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/revoke-token": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Çıkış",
        "description": "Çıkış",
        "operationId": "revokeTokenRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeTokenModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/get-all-accounts": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Sistemdeki Kayıtlı Hesaplar",
        "description": "Sistemdeki Kayıtlı hesaplar",
        "operationId": "getAllAccountsRequest",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customers/get-all": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Tüm Müşterileri Getirir",
        "description": "Tüm Müşterileri Getirir",
        "operationId": "findPetsByStatus",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountsResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/customers/get-customer/{id}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Müşteriyi Getirir",
        "description": "Müşteriyi Getirir",
        "operationId": "jdhfjdh",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/customers/add-customer": {
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Müşteri Ekler",
        "description": "Müşteri Ekler",
        "operationId": "findPetsByTags",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/delete-customer/{id}": {
      "delete": {
        "tags": [
          "Customers"
        ],
        "summary": "Müşteri Siler",
        "description": "Müşteri Siler",
        "operationId": "getPetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    },
    "/edit-customer/{id}": {
      "put": {
        "tags": [
          "Customers"
        ],
        "summary": "Müşteri Bilgilerini Düzenler",
        "description": "",
        "operationId": "uploadFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of pet to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "LoginModel": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "CustomerModel": {
        "type": "object",
        "properties": {
          "customerFirstName": {
            "type": "string"
          },
          "customerLastName": {
            "type": "string"
          },
          "customerCity": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "isVerified": {
            "type": "boolean"
          },
          "success": {
            "type": "boolean"
          },
          "credentials": {
            "type": "object"
          }
        }
      },
      "RegisterModel": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmPassword": {
            "type": "string"
          },
          "acceptTerms": {
            "type": "boolean"
          }
        }
      },
      "AccountsResponse": {
        "type": "object"
      },
      "CustomersResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "boolean"
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "number"
          }
        }
      },
      "RevokeTokenModel": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "RevokeTokenResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "requestBodies": {
      "Pet": {
        "description": "Pet object that needs to be added to the store",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginModel"
            }
          },
          "application/xml": {
            "schema": {
              "$ref": "#/components/schemas/LoginModel"
            }
          }
        }
      },
      "UserArray": {
        "description": "List of user object",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}